Setup beginnings fo typescript project
I set this up with a main entry point of main.ts as well as separate
ws.ts to be a typescript module to hold the webservice framework that is
going to be built as part of this design exploration.

I also setup jest so that I can use that to write automated tests to
drive this design exploration out in a provable way.

I wrapped most of the functionality for this up in the package.json
script sections and outlined most of them in the README.md as to how to
use them.