Add wf tests using handler builder pattern
I did this to see if using the handler function builder pattern would
help us with the dependency injection while still conforming to the
RouteHandler type. I also added the body property to the Request and
Response so that I could utilize them while testing to get unique
identifiable data out in tests.