Use the wf framework in manual test web app
I did this so that I could see and feel what it might look like to
actually build a web application with the framework. So this includes
starting to split things apart, e.g. splitting the routes builder out
into routes.ts and the various handlers out into their respective files,
etc.

I also modified the package.json so that the "run" script would properly
reference the ./build/src/main.js path rather than the old
./build/main.js path.

Not only does this show that using the framework thus far is reasonable
but it also shows via the main.test.ts that it is also possible to
facilitate dependency injection for end-to-end tests with the help of
the function builder pattern applied to the routes.